Refactored settings handling by DRY-ing the code.
Almost everything regarding options is now driven by an array of structs. The array not only specifies short and long option names and value requirements, but also descriptions and various flags. With this setup, it should only be necessary to setup options in a single place (the array) and the rest of the application should adapt accordingly. This greatly reduces the amount of work required when adding new options, for example adding them to cmd line parser, help or values print outs requires no manual work at all!

**NOTE:** Currently the only places where information needs to be repeated, is where option name must be used explicitly. For example when detecting whether `--version` or `--help` was issued or when specifying options that need to be treated as arrays. I left these cases for now, but if there would be more repetition, they can be easily updated in the future.