Fixed bug where method state expected body following method with arguments.
The problem was with a single BOOL set to YES. Have no idea how this happened. Prior unit tests didn't test for this so it wasn't detected.

Note that method state now also consumes semicolon after method definition. Although when used in actual application, this wouldn't be a problem due to Parser skipping over remnants, it's good practice to consume the whole thing when something is matched. I'd probably have to check for this in other parser states as well.