Added logging for matched constants.
Note that this "garbages" output when constants are detected in structs: it first spits out constant from the struct and then the struct itself. While this could be remedied with some flag indicating to constant state whether it should or not spit out the parsed data, I chose to leave it alone - this is internal logging only (kind of) that needs to be explicitly enabled. And it's more or less intended for debugging parsing.