Allow the selection of a template path at compile time
This path takes the least precedence, and allows packaging systems
like Homebrew to place the templates from the current version in
a path which this binary will look for as a last resort.

That makes the default case of, "please give me the latest version
of the templates," work seamlessly, while also allowing anyone to
override it using the existing instructions in the documentation.

--Tim Shadel <github@timshadel.com>