Refactored string normalization to work on whole words.
Note that this is somewhat less ideal as it only takes the first part of the string as opposed to prefix and suffix as was implemented before. Although it would be possible to implement whole word split at both sides, it would be a bit more complicated and involved. Another downside is missing separators and other nonstandard chars. However we're now using proper words boundaries that should work regardless of language of the represented text.

Will leave this code for the moment and deal with it later on in case I don't like the results.