Refactored process comment components unit tests to make failing cases more obvious.
Instead of comparing various components, we're now comparing primitives: strings and counts. The main pro of this approach is much more explicit case failure messages (i.e. expected A but got B). Another pro is there's no longer need to implement `toString:`. The downside is more verbose verification logic, but this should be overshadowed by the fact that's it much faster to see what exactly was wrong.