Implemented code block detection and storing.
Note that code blocks are treated the same as @warning or @bug - with special `CommentComponentInfo` subclass. The reason for such behavior is purely for making it easier for subsequent processing passes (such as detecting cross references) to skip code blocks entirely.

Also note that at this point detecting code block is implemented by checking for tab or four spaces prefix on each line. This is probably not the fastest implementation. Furthermore: the same detection is repeated for each detected code block - first when splitting whole string into sections and then when detecting types of components from sections. Oh, and also for most other components as this is the first test.