Added `--crossrefs` cmd line switch handling.
Note that the user can now supply not only the matcher regex, but also output format. The syntax is: `<matcher>:/<output>/`. Output format is optional, so `<matcher>` is also valid and will revert to default output format. However if output format is given, it must contain the following two placeholders: `$` and `%`. `$` is replaced by the cross reference description and `%` by actual cross reference link.