some clean-up for quick help (ALT+click) bubble fix in methods and setters.
If @property is used with setter= xxx and the setter's method is explicitly presented in @implementation section, quick help will link back to @property correctly.
But If @property is used with getter= xxx and the getter's method is explicitely presented in @implementation, the quick help on getter will not show @proprety's comments. A substantial modification of code is required to accommodate this functionality for custom getters. May fix it a bit later...