Separated finalizing a docset from installing it.
By finalizing, I mean delivering it to the install path. By installing it, I mean handing it to Xcode.

This distinction enables, for example, making an application that delivers a finished docset wherever the user wants it, as directed by a Save panel, without necessarily implicitly adding it to Xcode.