Fix headers in CommentsFormattingStyle.markdown
Github-flavored Markdown only recognizes headers if there's a space between the `#` characters and the text:

```
# a header
#not a header
```

This file was missing spaces on a bunch of headers so it didn't render correctly. This should fix it; if I missed any, it's probably easier to layer another patch on top of this one catching them rather than asking me to edit it, but whatever works!